Split off the integ and unit tests in different projects
