+++ SecretsManagerCachingDotNet.sln
@@ -9,13 +9,15 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{C6D686A7-A
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Amazon.SecretsManager.Extensions.Caching", "src\Amazon.SecretsManager.Extensions.Caching\Amazon.SecretsManager.Extensions.Caching.csproj", "{B7809806-7B35-4712-8995-30BB71218FE9}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Amazon.SecretsManager.Extensions.Caching.Tests", "test\Amazon.SecretsManager.Extensions.Caching.Tests\Amazon.SecretsManager.Extensions.Caching.Tests.csproj", "{72B442BD-CE77-4EF9-AF26-EBD4FD6B8972}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "solutionitems", "solutionitems", "{219B6F8C-9706-408E-8584-45045335DAAF}"
 	ProjectSection(SolutionItems) = preProject
 		.gitignore = .gitignore
 		code-analysis.ruleset = code-analysis.ruleset
 	EndProjectSection
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Amazon.SecretsManager.Extensions.Caching.UnitTests", "test\Amazon.SecretsManager.Extensions.Caching.UnitTests\Amazon.SecretsManager.Extensions.Caching.UnitTests.csproj", "{01186448-622D-4725-B21C-54A3FEBCB85F}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Amazon.SecretsManager.Extensions.Caching.IntegTests", "test\Amazon.SecretsManager.Extensions.Caching.IntegTests\Amazon.SecretsManager.Extensions.Caching.IntegTests.csproj", "{A07052B9-6D72-47FF-8EFD-6A35C58CF7FC}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -31,13 +33,22 @@ Global
 		{72B442BD-CE77-4EF9-AF26-EBD4FD6B8972}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{72B442BD-CE77-4EF9-AF26-EBD4FD6B8972}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{72B442BD-CE77-4EF9-AF26-EBD4FD6B8972}.Release|Any CPU.Build.0 = Release|Any CPU
+		{01186448-622D-4725-B21C-54A3FEBCB85F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{01186448-622D-4725-B21C-54A3FEBCB85F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{01186448-622D-4725-B21C-54A3FEBCB85F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{01186448-622D-4725-B21C-54A3FEBCB85F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A07052B9-6D72-47FF-8EFD-6A35C58CF7FC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A07052B9-6D72-47FF-8EFD-6A35C58CF7FC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A07052B9-6D72-47FF-8EFD-6A35C58CF7FC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A07052B9-6D72-47FF-8EFD-6A35C58CF7FC}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{B7809806-7B35-4712-8995-30BB71218FE9} = {73B3E414-F5D8-4495-A283-D4271126C53A}
-		{72B442BD-CE77-4EF9-AF26-EBD4FD6B8972} = {C6D686A7-A7EF-4396-9476-70A3A0FB9DEC}
+		{01186448-622D-4725-B21C-54A3FEBCB85F} = {C6D686A7-A7EF-4396-9476-70A3A0FB9DEC}
+		{A07052B9-6D72-47FF-8EFD-6A35C58CF7FC} = {C6D686A7-A7EF-4396-9476-70A3A0FB9DEC}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {1A5339B2-E624-44ED-B714-2F21D7406ED3}
